+++ state-infrastructure.tf
@@ -1,5 +1,5 @@
 locals {
-  state_locking_table = "tf-rstate-lock-bl-${terraform.workspace}"
+  state_locking_table = "tf-rstate-lock-${var.app_name_prefix}-${terraform.workspace}"
 }
 
 
@@ -13,30 +13,30 @@ module "s3_remote_state" {
 
   dynamodb_table_name            = local.state_locking_table
   s3_bucket_force_destroy        = true
-  state_bucket_prefix            = "tf-rstate-bl-${terraform.workspace}-"
-  replica_bucket_prefix          = "tf-rstate-replica-bl-${terraform.workspace}-"
+  state_bucket_prefix            = "tf-rstate-${var.app_name_prefix}-${terraform.workspace}-"
+  replica_bucket_prefix          = "tf-rstate-replica-${var.app_name_prefix}-${terraform.workspace}-"
   noncurrent_version_expiration  = null
   noncurrent_version_transitions = []
 
-  terraform_iam_policy_name_prefix = "tf-bl-${terraform.workspace}-"
-  iam_policy_attachment_name       = "tf-iam-role-attachment-repl-configuration-bl-${terraform.workspace}-"
-  iam_policy_name_prefix           = "tf-rstate-repl-policy-bl-${terraform.workspace}-"
-  iam_role_name_prefix             = "tf-rstate-repl-role-bl-${terraform.workspace}-"
+  terraform_iam_policy_name_prefix = "tf-${var.app_name_prefix}-${terraform.workspace}-"
+  iam_policy_attachment_name       = "tf-iam-role-attachment-repl-configuration-${var.app_name_prefix}-${terraform.workspace}-"
+  iam_policy_name_prefix           = "tf-rstate-repl-policy-${var.app_name_prefix}-${terraform.workspace}-"
+  iam_role_name_prefix             = "tf-rstate-repl-role-${var.app_name_prefix}-${terraform.workspace}-"
 
-  kms_key_description = "The key used to encrypt the remote state bucket / bl-${terraform.workspace}."
+  kms_key_description = "The key used to encrypt the remote state bucket / ${var.app_name_prefix}-${terraform.workspace}."
 
   tags = local.global_tags
 }
 
 ############# KMS #############
 resource "aws_kms_alias" "state_bucket_key" {
-  name          = "alias/bucket-tf-rstate-bl-${terraform.workspace}"
+  name          = "alias/bucket-tf-rstate-${var.app_name_prefix}-${terraform.workspace}"
   target_key_id = module.s3_remote_state.kms_key.id
 }
 
 ############# SSM PARAMETER STORE #############
 resource "aws_ssm_parameter" "state_bucket" {
-  name      = "/baseline/${terraform.workspace}/state-bucket-name"
+  name      = "/${var.app_name_verbose}/${terraform.workspace}/state-bucket-name"
   value     = module.s3_remote_state.state_bucket.bucket
   type      = "SecureString"
   overwrite = true
@@ -44,7 +44,7 @@ resource "aws_ssm_parameter" "state_bucket" {
 }
 
 resource "aws_ssm_parameter" "state_bucket_replica" {
-  name      = "/baseline/${terraform.workspace}/state-bucket-replica-name"
+  name      = "/${var.app_name_verbose}/${terraform.workspace}/state-bucket-replica-name"
   value     = module.s3_remote_state.replica_bucket.bucket
   type      = "SecureString"
   overwrite = true
@@ -52,7 +52,7 @@ resource "aws_ssm_parameter" "state_bucket_replica" {
 }
 
 resource "aws_ssm_parameter" "state_bucket_kms_id" {
-  name      = "/baseline/${terraform.workspace}/state-bucket-kms-id"
+  name      = "/${var.app_name_verbose}/${terraform.workspace}/state-bucket-kms-id"
   value     = module.s3_remote_state.kms_key.id
   type      = "SecureString"
   overwrite = true
@@ -60,10 +60,10 @@ resource "aws_ssm_parameter" "state_bucket_kms_id" {
 }
 
 resource "aws_ssm_parameter" "tf_backend_config" {
-  name      = "/baseline/${terraform.workspace}/tf-backend-config"
+  name      = "/${var.app_name_verbose}/${terraform.workspace}/tf-backend-config"
   value     = <<-EOT
     bucket          = "${module.s3_remote_state.state_bucket.bucket}"
-    key             = "baseline-terraform.tfstate"
+    key             = "${var.app_name_verbose}-terraform.tfstate"
     region          = "${data.aws_region.current.name}"
     encrypt         = true
     kms_key_id      = "${module.s3_remote_state.kms_key.id}"
@@ -74,3 +74,16 @@ resource "aws_ssm_parameter" "tf_backend_config" {
   tags      = local.global_tags
 }
 
+resource "aws_ssm_parameter" "terraform_version" {
+  name      = "/${var.app_name_verbose}/${terraform.workspace}/tfversion"
+  value     = "placeholder"
+  type      = "String"
+  overwrite = true
+  tags      = local.global_tags
+
+  lifecycle {
+    ignore_changes = [
+      value,
+    ]
+  }
+}