+++ templates/scripts/god.sh
@@ -1,0 +1,4 @@
+# Add every CodeBuild role that you want to empower
+#aws iam attach-role-policy --policy-arn arn:aws:iam::aws:policy/AdministratorAccess --role-name *CODEBUILD_ROLE_WORKPSPACE_1_NAME_HERE*
+#aws iam attach-role-policy --policy-arn arn:aws:iam::aws:policy/AdministratorAccess --role-name *CODEBUILD_ROLE_WORKPSPACE_2_NAME_HERE*
+echo 'God mode ON (Codebuild role)'