+++ ungod.sh
@@ -1,0 +1,2 @@
+echo "Setting Mortal mode"
+./private/scripts/ungod.sh