+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+## [0.3.0] - 2021-09-20
+### Added
+- Automated state location awarness
+- CI/CD pipeline in AWS (triggered by Githib events)
+- S3 bucket for CI/CD artifacti
+- IAM roles & policies for CodePipeline & CodeBuild
+- God mode for Codebuild triggered by developer with admin privileges
+
+### Changed
+- Substituted variables for app in prefix/verbose context
 
 ## [0.2.0] - 2021-09-10
 ### Added